fixing some compiler warnings, removing some unused lame_global_flags elements

+++ libmp3lame/util.c
@@ -20,7 +20,7 @@
  * Boston, MA 02111-1307, USA.
  */
 
-/* $Id: util.c,v 1.125 2006/09/10 14:25:35 bouvigne Exp $ */
+/* $Id: util.c,v 1.126 2006/11/07 01:35:19 robert Exp $ */
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
@@ -109,7 +109,7 @@ malloc_aligned(aligned_pointer_t * ptr, unsigned int size, unsigned int bytes)
         if (!ptr->pointer) {
             ptr->pointer = malloc(size + bytes);
             if (bytes > 0) {
-                ptr->aligned = (void *) ((((int) ptr->pointer + bytes - 1) / bytes) * bytes);
+                ptr->aligned = (void *) ((( (size_t)ptr->pointer + bytes - 1) / bytes) * bytes);
             }
             else {
                 ptr->aligned = ptr->pointer;
